transferred current self_supervised tests to integration
